fixed rankings and endpoints

+++ mongo/types.go
@@ -157,7 +157,7 @@ type UserRanking struct {
 	FID         uint64 `json:"fid" bson:"fid"`
 	Username    string `json:"username" bson:"username"`
 	Displayname string `json:"displayname" bson:"displayname"`
-	Count       uint64 `json:"count" bson:"count"`
+	Count       float64 `json:"count" bson:"count"`
 }
 
 // ElectionRanking is an election ranking entry.
